add azureite container with contents of data folder
connect api to azurite

add local env file

env setup for scheduled jobs

update data files to latest
