+++ .gitignore
@@ -1,5 +1,4 @@
 *.log
 __pycache__/
 *.pyc
-.env
-metadata.json
+.vscode/
