+++ api/app/logging_config.yaml
@@ -1,0 +1,41 @@
+version: 1
+disable_existing_loggers: false
+
+formatters:
+  default:
+    format: "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
+
+handlers:
+  console:
+    class: logging.StreamHandler
+    formatter: default
+    stream: ext://sys.stdout
+  
+  main_file:
+    class: logging.handlers.RotatingFileHandler
+    formatter: default
+    filename: logs/main.log
+    maxBytes: 10485760  # 10MB
+    backupCount: 5
+  
+  search_engine_file:
+    class: logging.handlers.RotatingFileHandler
+    formatter: default
+    filename: logs/search_engine.log
+    maxBytes: 10485760
+    backupCount: 5
+
+loggers:
+  main:
+    level: INFO
+    handlers: [console, main_file]
+    propagate: false
+  
+  search_engine:
+    level: INFO
+    handlers: [console, search_engine_file]
+    propagate: false
+
+root:
+  level: INFO
+  handlers: [console, main_file]
