+++ api/docker-compose.yml
@@ -1,0 +1,13 @@
+version: '3.8'
+
+services:
+  fastapi_app:
+    build:
+      context: .
+    volumes:
+      - ./app:/api/app
+      - ../data:/api/data  # Mount /epiverse-search-backend/data to /api/app/data
+    ports:
+      - "8000:8000"  # Map container port 8000 to host port 8000
+    env_file:
+      - .env
