--- docker-compose.yml
@@ -1,31 +1,0 @@
-version: '3.8'
-
-services:
-
-#    loader:
-#        build:
-#            context: .
-#            dockerfile: Dockerfile
-#        container_name: loader
-#        depends_on:
-#            - elasticsearch
-
-
-    backend:
-        image : pytorch/pytorch
-        build:
-          context: .
-          dockerfile: Dockerfile
-#          platforms: 
-#            - linux/arm64
-        container_name: fastapi_cross_encoder_20250117
-#        depends_on:
-#            - elasticsearch
-        
-        ports:
-            - "80:80"
-#        extra_hosts:
-#            - "host.docker.internal:8000"
-        command: ["uvicorn", "search_api_endpoint:app", "--host", "0.0.0.0", "--port", "80", "--reload"]
-
-
