+++ jobs/fetch-universe-metadata/docker-compose.yml
@@ -4,29 +4,11 @@ services:
     
     volumes:
       - .:/home/site/wwwroot  # Mount code into container, not needed for production
-    
-    depends_on:
-      - azurite 
+
     ports:
       - "7071:80"
-    environment:
-      - AzureWebJobsStorage=DefaultEndpointsProtocol=http;AccountName=devstoreaccount1;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;BlobEndpoint=http://azurite:10000/devstoreaccount1;QueueEndpoint=http://azurite:10001/devstoreaccount1;TableEndpoint=http://azurite:10002/devstoreaccount1;
+    env_file:
+      - .env
     
     entrypoint: ["func", "host", "start", "--verbose"]
 
-
-  azurite:
-    image: mcr.microsoft.com/azure-storage/azurite
-    container_name: azurite
- 
-    ports:
-      - "10000:10000"  # Blob Storage
-      - "10001:10001"  # Queue Storage
-      - "10002:10002"  # Table Storage
-    volumes:
-      - azurite_data:/data  # Persist storage  
-    command: "azurite --loose --blobHost 0.0.0.0 --queueHost 0.0.0.0 --tableHost 0.0.0.0"
-
-    
-volumes:
-  azurite_data:
